+++ config/app.php
@@ -24,7 +24,7 @@ return [
     |
     */
 
-    'version' => '2.0.0',
+    'version' => '2.1.0',
 
     /*
     |--------------------------------------------------------------------------
@@ -177,6 +177,7 @@ return [
          * Package Service Providers...
          */
         Collective\Html\HtmlServiceProvider::class, // Forms & HTML
+        SocialiteProviders\Manager\ServiceProvider::class, // Socialite Providers
 
         /*
          * Application Service Providers...
@@ -244,6 +245,7 @@ return [
          */
         'Form' => Collective\Html\FormFacade::class, // Forms
         'Html' => Collective\Html\HtmlFacade::class, // HTML
+        'Socialite' => Laravel\Socialite\Facades\Socialite::class, // Socialite
 
     ],