+++ database/migrations/2020_03_25_114812_prepare_users_table_for_social_authentication.php
@@ -1,0 +1,48 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Schema;
+
+class PrepareUsersTableForSocialAuthentication extends Migration
+{
+    /**
+     * PrepareUsersTableForSocialAuthentication constructor.
+     *
+     * Forces Doctrine to treat enum columns as string.
+     */
+    public function __construct()
+    {
+        DB::getDoctrineSchemaManager()
+            ->getDatabasePlatform()
+            ->registerDoctrineTypeMapping('enum', 'string');
+    }
+
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('users', function (Blueprint $table) {
+            // Making email and password nullable
+            $table->string('email')->nullable()->change();
+            $table->string('password')->nullable()->change();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('users', function (Blueprint $table) {
+            $table->string('email')->nullable(false)->change();
+            $table->string('password')->nullable(false)->change();
+        });
+    }
+}