+++ resources/lang/en/social_login.php
@@ -1,0 +1,21 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Social login
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are used during authentication for various
+    | messages that we need to display to the user. You are free to modify
+    | these language lines according to your application's requirements.
+    |
+    */
+
+    'okta' => 'Sign in using Okta',
+    'github' => 'Sign in using Github',
+    'facebook' => 'Sign in using Facebook',
+    'e-mail' => 'Sign in using e-mail',
+
+];